stockbot/requirements.txt
discord.py>=2.3.2
fastapi>=0.109.0
uvicorn>=0.27.0
playwright>=1.41.0
redis>=5.0.1
apscheduler>=3.10.4
python-dotenv>=1.0.0
gunicorn>=21.2.0
jinja2>=3.1.3
pydantic>=2.5.3
pydantic-settings>=2.1.0
regex>=2023.12.25
aiohttp>=3.9.1
